Functionality to build s2e and run binaries
Added commands in the justfile for:
 * Downloading s2e source
 * Integrating BinRec-plugins into the s2e-tree
 * Building s2e
 * Rebuilding s2e-plugins
 * Create new analysis projects
 * Running analysis projects

Creating new analysis projects and running them is implemented in
python.

This commit probably breaks the functionality due to changing the
S2E environment variable. That will be fixed in a following commit.

There is still some cleanup needed. Specifically, comments and test
cases for the python code.